+++ src/app/docs/components/tabs/page.mdx
@@ -1,7 +1,7 @@
-import SimpleTabs from '@/showcase/components/SimpleTabs'
-import ButtonShapeTabs from '@/showcase/components/ButtonShapeTabs'
-import LineTabs from '@/showcase/components/LineTabs'
-import IconTabs from '@/showcase/components/IconTabs'
+import SimpleTabs from '@/showcase/components/tabs/SimpleTabs'
+import ButtonShapeTabs from '@/showcase/components/tabs/ButtonShapeTabs'
+import LineTabs from '@/showcase/components/tabs/LineTabs'
+import IconTabs from '@/showcase/components/tabs/IconTabs'
 import PreviewCodeTabs from '@/components/CodePreview/page'
 
 export const metadata = {
