+++ src/showcase/components/tabs/SimpleTabs.tsx
@@ -1,19 +1,19 @@
-'use client'
+'use client';
 
-import { Tabs } from '@/ui/Tabs'
-import { useState } from 'react'
+import { Tabs } from '@/showcase/components/tabs/Tabs';
+import { useState } from 'react';
 
 const SimpleTabs = () => {
-  const [activeTab, setActiveTab] = useState('ansub')
+  const [activeTab, setActiveTab] = useState('ansub');
 
   const tabs = [
     { id: 'ansub', label: 'Ansub' },
     { id: 'syntax', label: 'Syntax UI' },
-  ]
+  ];
 
   const handleTabClick = (tabId: string) => {
-    setActiveTab(tabId)
-  }
+    setActiveTab(tabId);
+  };
 
   return (
     <div>
@@ -23,7 +23,7 @@ const SimpleTabs = () => {
         <div>syntaxUI is for people who want to start rapid development.</div>
       )}
     </div>
-  )
-}
+  );
+};
 
-export default SimpleTabs
+export default SimpleTabs;
